Fixed compilation problems

+++ source/vehicle.cpp
@@ -432,7 +432,7 @@ void Vehicle :: setNewHeight( int newHeight )
   if ( maxMovement() ) {
      float oldperc = float(getMovement ( false )) / float(maxMovement());
      height = newHeight;
-     setMovement ( floor(maxMovement() * oldperc + 0.5) , 0 );
+     setMovement ( int(floor(maxMovement() * oldperc + 0.5)) , 0 );
   } else {
      height = newHeight;
      warning("Internal error: unit has invalid height");
